+++ app/src/main/java/ru/studyguk/cryptoapp/data/database/CoinInfoDbModel.kt
@@ -13,5 +13,5 @@ data class CoinInfoDbModel(
     val highDay: String?,
     val lowDay: String?,
     val lastMarket: String?,
-    val imageUrl: String?
+    val imageUrl: String
 )
