+++ app/src/main/java/ru/studyguk/cryptoapp/data/network/ApiFactory.kt
@@ -5,7 +5,6 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 object ApiFactory {
     private const val BASE_URL = "https://min-api.cryptocompare.com/data/"
-    const val BASE_IMAGE_URL = "https://www.cryptocompare.com"
 
     private val retrofit = Retrofit.Builder()
         .addConverterFactory(GsonConverterFactory.create())
