+++ app/src/main/java/ru/studyguk/cryptoapp/domain/CoinInfo.kt
@@ -4,9 +4,9 @@ data class CoinInfo(
     val fromSymbol: String,
     val toSymbol: String?,
     val price: String?,
-    val lastUpdate: Long?,
+    val lastUpdate: String,
     val highDay: String?,
     val lowDay: String?,
     val lastMarket: String?,
-    val imageUrl: String?
+    val imageUrl: String
 )
