--- app/src/main/java/ru/studyguk/cryptoapp/utils/TimeUtils.kt
@@ -1,15 +1,0 @@
-package ru.studyguk.cryptoapp.utils
-
-import java.sql.Timestamp
-import java.text.SimpleDateFormat
-import java.util.*
-
-fun convertTimeStampToTime(timestamp: Long?): String {
-    if (timestamp == null) return ""
-    val stamp = Timestamp(timestamp * 1000)
-    val date = Date(stamp.time)
-    val pattern = "HH:mm:ss"
-    val sdf = SimpleDateFormat(pattern, Locale.getDefault())
-    sdf.timeZone = TimeZone.getDefault()
-    return sdf.format(date)
-}
